requirements.txt
asgiref==3.11.1
Django==6.0.2
django-filter==25.2
django-simple-history==3.11.0
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==26.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
rest_framework_simplejwt